routes: migrate health domain (9 routes)
Move to src/server/routes/health.ts:
- GET  /api/health
- POST /api/internal/test/replay-buffered-events/:sessionId
- GET  /api/setup-status
- POST /api/setup-status/dismiss
- GET  /api/system-prompt-context
- PUT  /api/system-prompt-context
- POST /api/system-prompt/customise
- POST /api/shutdown
- GET  /api/ca-cert

Routes registered in dispatcher; legacy handleApiRoute branches removed.

Net: server.ts -158 lines (7922 → 7764). No behaviour change.
