+++ src/server/server.ts
@@ -18,7 +18,6 @@ import { RateLimiter } from "./auth/rate-limit.js";
 import { validateToken } from "./auth/token.js";
 import { oauthComplete, oauthFlowStatus, oauthStart, oauthStatus } from "./auth/oauth.js";
 import { handleWebSocketConnection } from "./ws/handler.js";
-import { paceAndSend, PACE_TIMEOUT_MS } from "./replay-pacing.js";
 import { discoverSlashSkills, getSkillDirectories, getSlashSkill, buildSlashSkillPrompt } from "./skills/slash-skills.js";
 import { TeamManager, GateDependencyError } from "./agent/team-manager.js";
 import { checkGateDependencies } from "./agent/gate-dependency-check.js";
@@ -1160,163 +1159,6 @@ async function handleApiRoute(
 			if (task) return getTaskManagerForGoal(task.goalId);
 		}
 		throw new Error(`Task "${taskId}" not found in any project`);
-	}
-
-	// GET /api/health — unauthenticated so the client can probe localhost mode
-	if (url.pathname === "/api/health" && req.method === "GET") {
-		const isLocalhost = !config.forceAuth && (config.host === "localhost" || config.host === "127.0.0.1" || config.host === "::1");
-		json({
-			status: "ok",
-			sessions: sessionManager.listSessions().length,
-			localhost: isLocalhost,
-			aigw: !!getAigwUrl(preferencesStore),
-			setupComplete: isSetupComplete(),
-			orphanedTranscripts: sessionManager.orphanedTranscriptsCount,
-		});
-		return;
-	}
-
-	// POST /api/internal/test/replay-buffered-events/:sessionId — BOBBIT_E2E-only hook
-	// used by ST-DEDUP-01 to reproduce live-streaming duplication. Iterates the
-	// session's EventBuffer and re-broadcasts each buffered entry on the SAME
-	// wire path production uses, so when the fix adds `seq`/`ts` to the
-	// broadcast envelope the endpoint will naturally carry them too (because it
-	// inspects the buffer's stored entries — which upgrade from raw events to
-	// {seq,ts,event} tuples post-fix). Pre-fix: clients receive duplicate
-	// events and dupe-append assistant/toolResult messages. Post-fix: clients
-	// dedupe by seq and the message list stays stable.
-	const replayMatch = url.pathname.match(/^\/api\/internal\/test\/replay-buffered-events\/([^/]+)$/);
-	if (replayMatch && req.method === "POST") {
-		if (process.env.BOBBIT_E2E !== "1") { json({ error: "BOBBIT_E2E not enabled" }, 403); return; }
-		const sessionId = replayMatch[1];
-		const session = sessionManager.getSession(sessionId);
-		if (!session) { json({ error: "session not found" }, 404); return; }
-		const entries = session.eventBuffer.getAll() as any[];
-		let replayed = 0;
-		const deadline = Date.now() + PACE_TIMEOUT_MS;
-		for (const entry of entries) {
-			// Accept both raw-event shape (pre-fix) and {seq,ts,event} (post-fix).
-			const isWrapped = entry && typeof entry === "object" && "event" in entry && ("seq" in entry || "ts" in entry);
-			const framePayload = isWrapped
-				? { type: "event" as const, data: entry.event, seq: entry.seq, ts: entry.ts }
-				: { type: "event" as const, data: entry };
-			const data = JSON.stringify(framePayload);
-			for (const client of session.clients) {
-				await paceAndSend(client as any, data, deadline);
-			}
-			replayed++;
-		}
-		json({ replayed, bufferSize: session.eventBuffer.size });
-		return;
-	}
-
-	// GET /api/setup-status — check if project setup has been completed
-	if (url.pathname === "/api/setup-status" && req.method === "GET") {
-		json({ complete: isSetupComplete() });
-		return;
-	}
-
-	// POST /api/setup-status/dismiss — mark setup as dismissed (writes sentinel file)
-	if (url.pathname === "/api/setup-status/dismiss" && req.method === "POST") {
-		const stateDir = bobbitStateDir();
-		fs.mkdirSync(stateDir, { recursive: true });
-		fs.writeFileSync(path.join(stateDir, "setup-complete"), "dismissed\n");
-		json({ ok: true });
-		return;
-	}
-
-	// GET /api/system-prompt-context — read the project context section from system-prompt.md
-	if (url.pathname === "/api/system-prompt-context" && req.method === "GET") {
-		const systemPromptPath = path.join(bobbitConfigDir(), "system-prompt.md");
-		if (!fs.existsSync(systemPromptPath)) { json({ context: "" }); return; }
-		try {
-			const content = fs.readFileSync(systemPromptPath, "utf-8");
-			// Extract everything after the last "# Project Context" heading, or return empty
-			const marker = "# Project Context";
-			const idx = content.lastIndexOf(marker);
-			if (idx === -1) { json({ context: "" }); return; }
-			const context = content.slice(idx + marker.length).trim();
-			json({ context });
-		} catch { json({ context: "" }); }
-		return;
-	}
-
-	// PUT /api/system-prompt-context — append/replace the project context section in system-prompt.md
-	if (url.pathname === "/api/system-prompt-context" && req.method === "PUT") {
-		const body = await readBody(req);
-		if (!body || typeof body.context !== "string") { json({ error: "Missing context" }, 400); return; }
-		const systemPromptPath = path.join(bobbitConfigDir(), "system-prompt.md");
-		try {
-			let existing = "";
-			if (fs.existsSync(systemPromptPath)) {
-				existing = fs.readFileSync(systemPromptPath, "utf-8");
-			}
-			const marker = "# Project Context";
-			const idx = existing.lastIndexOf(marker);
-			const base = idx !== -1 ? existing.slice(0, idx).trimEnd() : existing.trimEnd();
-			const newContent = base + "\n\n" + marker + "\n\n" + body.context.trim() + "\n";
-			fs.mkdirSync(path.dirname(systemPromptPath), { recursive: true });
-			fs.writeFileSync(systemPromptPath, newContent);
-			json({ ok: true });
-		} catch (err: any) {
-			jsonError(500, err);
-		}
-		return;
-	}
-
-	// POST /api/system-prompt/customise — copy shipped default to .bobbit/config/system-prompt.md
-	//   so the user can edit it. If the file already exists it is left unchanged.
-	//   Returns { path, created, content }.
-	if (url.pathname === "/api/system-prompt/customise" && req.method === "POST") {
-		const userPath = path.join(bobbitConfigDir(), "system-prompt.md");
-		const defaultPath = path.join(
-			path.dirname(fileURLToPath(import.meta.url)),
-			"defaults",
-			"system-prompt.md",
-		);
-		let created = false;
-		try {
-			if (!fs.existsSync(userPath)) {
-				if (!fs.existsSync(defaultPath)) {
-					json({ error: "Default system-prompt.md not found in install" }, 500);
-					return;
-				}
-				fs.mkdirSync(path.dirname(userPath), { recursive: true });
-				fs.copyFileSync(defaultPath, userPath);
-				created = true;
-			}
-			const content = fs.readFileSync(userPath, "utf-8");
-			json({ path: userPath, created, content });
-		} catch (err: any) {
-			jsonError(500, err);
-		}
-		return;
-	}
-
-	// POST /api/shutdown — graceful shutdown (used by coverage teardown to flush V8 coverage)
-	if (url.pathname === "/api/shutdown" && req.method === "POST") {
-		json({ status: "shutting down" });
-		// Defer exit to allow the response to be sent
-		setTimeout(() => process.exit(0), 500);
-		return;
-	}
-
-	// GET /api/ca-cert — download the Bobbit CA certificate for device trust
-	if (url.pathname === "/api/ca-cert" && req.method === "GET") {
-		const caCertPath = config.tls?.caCert;
-		if (!caCertPath || !fs.existsSync(caCertPath)) {
-			json({ error: "No CA certificate available. Server is using a self-signed certificate." }, 404);
-			return;
-		}
-		const certData = fs.readFileSync(caCertPath);
-		res.writeHead(200, {
-			// iOS Safari needs this MIME type to offer the profile-install flow.
-			"Content-Type": "application/x-x509-ca-cert",
-			"Content-Disposition": "attachment; filename=\"bobbit-ca.crt\"",
-			"Content-Length": certData.length,
-		});
-		res.end(certData);
-		return;
 	}
 
 	// GET /api/sandbox-pool (deprecated — no longer a real pool, returns basic stats)
